main: common leds api color format
